Add reaction handlers to respond to user actions
