+++ bot/src/reactions/index.js
@@ -1,0 +1,9 @@
+const { northWestArrow } = require("./north-west-arrow");
+const { southEastArrow } = require("./south-east-arrow");
+const { crossMark } = require("./cross-mark");
+
+module.exports = {
+  "↖": northWestArrow,
+  "↘": southEastArrow,
+  "❌": crossMark
+};
